requirements.txt
# for dumping
pyserial

# for jupyter visualization
jupyter
matplotlib
numpy
scipy
